# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BlindPass SPS API
  version: 0.1.0
  description: |
    Maintained OpenAPI snapshot for the stable BlindPass hosted and local-development API surfaces.
    This document is derived from the route schemas in `packages/sps-server/src/routes`.
servers:
  - url: http://127.0.0.1:3100
    description: Local development
  - url: https://sps.atas.tech
    description: Hosted production
...
paths:
...
  /api/v2/dashboard/summary:
    get:
      tags: [Billing]
      summary: Dashboard overview model
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Workspace billing and quota summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
  /api/v2/billing:
    get:
      tags: [Billing]
      summary: Read the current workspace billing record
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Billing record
          content:
            application/json:
              schema:
                type: object
                properties:
                  billing:
                    $ref: '#/components/schemas/BillingRecord'
                required: [billing]
  /api/v2/billing/checkout:
    post:
      tags: [Billing]
      summary: Create a hosted billing checkout session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url:
                    type: string
                    format: uri
                  session_id:
                    type: string
                  billing:
                    $ref: '#/components/schemas/BillingRecord'
                required: [checkout_url, session_id, billing]
  /api/v2/billing/portal:
    post:
      tags: [Billing]
      summary: Create a hosted billing portal session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Portal session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  portal_url:
                    type: string
                    format: uri
                required: [portal_url]
  /api/v2/billing/allowances:
    get:
      tags: [Billing]
      summary: List per-agent x402 allowances
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Allowance list
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowances:
                    type: array
                    items:
                      $ref: '#/components/schemas/BillingAllowance'
                required: [allowances]
    post:
      tags: [Billing]
      summary: Create or update an agent x402 allowance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [agent_id, monthly_budget_cents]
              properties:
                agent_id:
                  type: string
                monthly_budget_cents:
                  type: integer
                  minimum: 0
      responses:
        '201':
          description: Allowance upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  allowance:
                    $ref: '#/components/schemas/BillingAllowance'
                required: [allowance]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentApiKey:
      type: apiKey
      in: header
      name: x-agent-api-key
...
  schemas:
...
    BillingRecord:
      type: object
      properties:
        workspace_id: { type: string }
        workspace_slug: { type: string }
        tier: { type: string }
        status: { type: string }
        billing_provider: { type: string, nullable: true }
        provider_customer_id: { type: string, nullable: true }
        provider_subscription_id: { type: string, nullable: true }
        subscription_status: { type: string }
      required: [workspace_id, workspace_slug, tier, status, billing_provider, provider_customer_id, provider_subscription_id, subscription_status]
    DashboardSummary:
      type: object
      properties:
        workspace:
          type: object
          properties:
            id: { type: string }
            slug: { type: string }
            display_name: { type: string }
            tier: { type: string }
            status: { type: string }
          required: [id, slug, display_name, tier, status]
        billing:
          $ref: '#/components/schemas/BillingRecord'
        counts:
          type: object
          properties:
            active_agents: { type: integer }
            active_members: { type: integer }
          required: [active_agents, active_members]
        quota:
          type: object
          properties:
            secret_requests:
              type: object
              properties:
                used: { type: integer }
                limit: { type: integer }
                reset_at: { type: integer }
              required: [used, limit, reset_at]
            agents:
              type: object
              properties:
                used: { type: integer }
                limit: { type: integer }
              required: [used, limit]
            members:
              type: object
              properties:
                used: { type: integer }
                limit: { type: integer }
              required: [used, limit]
            a2a_exchange_available:
              type: boolean
          required: [secret_requests, agents, members, a2a_exchange_available]
      required: [workspace, billing, counts, quota]
    BillingAllowance:
      type: object
      properties:
        agent_id: { type: string }
        display_name: { type: string, nullable: true }
        status: { type: string, nullable: true }
        monthly_budget_cents: { type: integer }
        current_spend_cents: { type: integer }
        remaining_budget_cents: { type: integer }
        budget_reset_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [agent_id, display_name, status, monthly_budget_cents, current_spend_cents, remaining_budget_cents, budget_reset_at, updated_at]
...